+++ sdk/storage/Azure.Storage.DataMovement/src/TransferProgressTracker.cs
@@ -12,7 +12,7 @@ namespace Azure.Storage.DataMovement
     /// Internal class for tracking progress of transfers and reporting back to
     /// user's progress handler.
     /// </summary>
-    internal class TransferProgressTracker : IAsyncDisposable
+    internal class TransferProgressTracker
     {
         internal class ProgressEventArgs
         {
@@ -48,10 +48,10 @@ namespace Azure.Storage.DataMovement
             _progressProcessor.Process = ProcessProgressEvent;
         }
 
-        public async ValueTask DisposeAsync()
+        public async Task TryCompleteAsync()
         {
             // This will close the channel and block on all pending items being processed
-            await _progressProcessor.DisposeAsync().ConfigureAwait(false);
+            await _progressProcessor.TryCompleteAsync().ConfigureAwait(false);
         }
 
         public async ValueTask IncrementCompletedFilesAsync(CancellationToken cancellationToken)